HPCC-19048 Unable to refresh locked session
When a user locks a session and tries to open a new tab or refreshes the locked screen they will get sent to an eclwatch without any
information being sent from the server. ESP attempts to redirect but the browser
has cached the content and is stuck in a loop.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>